Create rugbycalculatorGC23
Creating a scoring system for Rugby. (Rugby Union or league, your choice) This program asks for the two team names and each Try, conversion & penalty kick is to be recorded separately, so that at the end of the game. Not only do we know the scores for the 2 teams but also how the points were scored. (25 points)